+++ changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.1.2] - 2024-10-25
+
+### Changed
+
+- Modify `GameProgram` struct to pass all types of events to `input` instead of just `WindowEvent`s, but to intercept and filter out events that are useful for resizing etc.
+- Modify `Game` trait `resize` and `input` functions to take a reference instead of an owned value.
+- Rename `RenderGame` to `ExampleGame` in examples.
+
 ## [0.1.1] - 2024-10-09
 
 ### Added